requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
orjson
python-dotenv

langgraph
langchain
langchain-openai
langchain-google-genai
langchain-ollama
langchain-chroma

chromadb

jinja2
pandas
tabulate
python-dateutil
aiosqlite

tqdm
tenacity
rich

openai
trafilatura
